Moving go server into its own sub dir
