+++ backend/base_manager.go
@@ -1,4 +1,4 @@
-package main
+package typingwars
 
 import "log"
