+++ backend/event_dispatcher.go
@@ -1,4 +1,4 @@
-package main
+package typingwars
 
 import (
 	"time"
