+++ backend/game_over.go
@@ -1,4 +1,4 @@
-package main
+package typingwars
 
 import "github.com/gofrs/uuid"
