+++ backend/server.go
@@ -1,4 +1,4 @@
-package main
+package typingwars
 
 import (
 	"html/template"
