+++ backend/time_tick.go
@@ -1,4 +1,4 @@
-package main
+package typingwars
 
 type TimeTick struct {
 	FrameID uint32
