+++ backend/unit_collision.go
@@ -1,4 +1,4 @@
-package main
+package typingwars
 
 type UnitCollision struct {
 	Unit *Unit
