+++ backend/unit_spawned.go
@@ -1,4 +1,4 @@
-package main
+package typingwars
 
 type UnitSpawned struct {
 	Unit *Unit
